clean up identity aggregate

+++ Backend/Services/IdentityService/IdentityService.Application/CommandHandlers/Admins/RemoveRoleFromUserCommandHandler.cs
@@ -1,7 +1,11 @@
-﻿using IdentityService.Application.Commands.Admins;
+﻿using BuildingBlocks.Exceptions.Domain;
+using IdentityService.Application.Commands.Admins;
 using IdentityService.Application.Services;
-using IdentityService.Core.Entities;
+using IdentityService.Core.Aggregates;
+using IdentityService.Core.Identities;
+using IdentityService.Core.Models;
 using MediatR;
+using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -10,18 +14,39 @@ using System.Threading.Tasks;
 
 namespace IdentityService.Application.CommandHandlers.Admins
 {
-    public class RemoveRoleFromUserCommandHandler : IRequestHandler<RemoveRoleFromUserCommand, ApplicationUser>
+    public class RemoveRoleFromUserCommandHandler : IRequestHandler<RemoveRoleFromUserCommand, InternalUserModel>
     {
+        private readonly IApplicationUserRepository _applicationUserRepository;
         private readonly IAdminService _adminService;
+        private readonly UserManager<InternalIdentityUser> _userManager;
 
-        public RemoveRoleFromUserCommandHandler(IAdminService adminService)
+        public RemoveRoleFromUserCommandHandler(IApplicationUserRepository applicationUserRepository, IAdminService adminService, UserManager<InternalIdentityUser> userManager)
         {
+            _applicationUserRepository = applicationUserRepository;
             _adminService = adminService;
+            _userManager = userManager;
         }
 
-        public async Task<ApplicationUser> Handle(RemoveRoleFromUserCommand request, CancellationToken cancellationToken)
+        public async Task<InternalUserModel> Handle(RemoveRoleFromUserCommand request, CancellationToken cancellationToken)
         {
-            return await _adminService.RemoveRoleFromUser(request.UserId, request.Roles);
+            var applicationUser = await _applicationUserRepository.FindAsync(request.UserId);
+
+            if (applicationUser == null)
+            {
+                throw new AggregateNotFoundException(nameof(ApplicationUser), request.UserId);
+            }
+
+            var identityUser = await _adminService.RemoveRoleFromUser(request.UserId, request.Roles);
+
+            var roles = await _userManager.GetRolesAsync(identityUser);
+
+
+            return new InternalUserModel
+            {
+                ApplicationUser = applicationUser,
+                InternalIdentityUser = identityUser,
+                Roles = roles.ToList()
+            };
         }
     }
 }
